ARMv7-A (AArch32) port of KernelPatch
Port scope (35 adapted files + 15 new files):
- kernel/arm/                      AArch32 arch support (5 files)
- kernel/linux/arch/arm/           ported linux arch/arm headers (9 files)
- kernel/Makefile.arm              A32 cross build (arm-linux-gnueabi-gcc, -marm, no LPAE)
- kernel/kpimg-arm{,-full}.lds     A32 linker scripts for the payload
- kernel/base/setup_arm.S, kernel/base/puff/setjmp_arm.S
- kernel/include/kp_visibility_arm.h, kernel/check_struct.c
- adapted: base/{hook,hotpatch,fphook,map,start}.c, include/{pgtable,log,io,cache,
  barrier,hook,stdint}.h, patch/{patch.c,common/*,ksyms/*,include/*},
  tools/{image,symbol,kallsym,patch,common}.{c,h}
- patch/module/* deliberately left AArch64: the KPM .ko loader is NOT ported yet
  (not needed for the APatch/kernel-root path)

Verified on the QEMU oracle (linux 5.15.167, armv7, -M virt -cpu cortex-a15):
- gate A (clean full rebuild -> link -> kptools -> zImage -> boot):
  0 panic / 0 fault / 0 Internal error, userspace reached 1, initmem freed 1,
  clean poweroff 1, KP start() ran (KP Restore: skipped), kernel at link PA 40208000
- gate B (kernel-side escalation, end to end):
  supercall HELLO = 0x11581158 (magic ok), SUPERCALL_SU rc 0, uid 1000 -> 0 via hook

Verified against the real device kernel (XTC Z9, msm8937, 4.9.193-perf, AArch32):
- kptools -p exit 0, static self-checks pass, bss extent 0x1889a84,
  .setup.map search picks 0x13f7000..0x13f8000 (symbol free, zero, RWX half)
- device boot of a patched image is NOT yet proven (see port/XTC_BOOT_REPACK_REPORT.md)

Verified payload (not committed, rebuild with `make -f kernel/Makefile.arm full`):
  kernel/kpimg-arm-full  252076 bytes  md5 556296d34677cb1509db9b64b7b2b09d

+++ kernel/patch/common/accctl.c
@@ -243,7 +243,7 @@ int bypass_selinux()
     if (avc_denied_addr) {
         hook_err_t err = hook((void *)avc_denied_addr, (void *)avc_denied_replace, (void **)&avc_denied_backup);
         if (err != HOOK_NO_ERR) {
-            log_boot("hook avc_denied_addr: %llx, error: %d\n", avc_denied_addr, err);
+            log_boot("hook avc_denied_addr: %llx, error: %d\n", (unsigned long long)avc_denied_addr, err);
         }
     }
 
@@ -252,7 +252,7 @@ int bypass_selinux()
         hook_err_t err =
             hook((void *)slow_avc_audit_addr, (void *)slow_avc_audit_replace, (void **)&slow_avc_audit_backup);
         if (err != HOOK_NO_ERR) {
-            log_boot("hook slow_avc_audit: %llx, error: %d\n", slow_avc_audit_addr, err);
+            log_boot("hook slow_avc_audit: %llx, error: %d\n", (unsigned long long)slow_avc_audit_addr, err);
         }
     }
